refactor(mina.ts, transaction.ts): move transaction function from mina.ts to transaction.ts
